requirements.txt

Django==4.2.3
django-cors-headers==4.2.0
django-cron==0.6
django-environ==0.10.0
django-extensions==3.2.3
django-import-export==3.2.0

django-import-export==3.2.0

requests==2.31.0


pymysql==1.1.0
python-dotenv==1.0.0
django-simple-history==3.3.0
#  used for livekit
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
drf-api-tracking==1.8.0
drf-spectacular==0.26.3
drf-writable-nested==0.7.0


